# strings, builtin: remove `strings.Builder.clear()`, fix `array.clear()` not working in the JS backend