<!--

Please title your PR as follows: `module: description` (e.g. `time: fix date format`).
Always start with the thing you are fixing, then describe the fix.
Don't use past tense (e.g. "fixed foo bar").

Explain what your PR does and why.

If you are adding a new function, please document it and add tests:

```
// foo does foo and bar
fn foo() {

// file_test.v
fn test_foo() {
    assert foo() == ...
    ...
}
```

If you are fixing a bug, please add a test that covers it.

Before submitting a PR, please run `v test-all` .
See also `TESTS.md`.

I try to process PRs as soon as possible. They should be handled within 24 hours.

Applying labels to PRs is not needed.

Thanks a lot for your contribution!

-->

Currently, `strings.Builder` has a function named `clear()` which clears the content by reallocating the backing array.

However, there is no need to reallocate the array. For example, in `str()` function, the backing array is simply trimmed to zero length.

Also, the array type itself has a `clear()` function which does the same as `trim(0)`. This PR reuses the `clear()` function in the `array` by simply removing the `clear()` function in `strings.Builder` and replacing `trim(0)` with `clear()` which does the same thing.

This PR shouldn't break existing code