@XiaoPangxie732 2025-03-21
There seems to be no `clear()` method in JS `Array`s, so I use the approach grabbed from the internet