@spytheman 2025-03-21
What was the initial motivation for this change?
Code cleanup, performance improvement, or a bugfix?