@XiaoPangxie732 2025-03-22
> What was the initial motivation for this change?
> 
> Code cleanup, performance improvement, or a bugfix?

Initially it was a performance improvement, and later I thought it would be more readable to use `clear()` than `trim(0)` and could avoid a `index < a.len` check. But CI then failed. I found that the reason is that `array.clear()` is badly implemented on the js backend, but I don't know much about js, so I just searched a way to clear it on the web, and I'm not sure if there's a better way to do it, sorry